+++ RELEASE_NOTES.md
@@ -9,3 +9,9 @@ Until then, breaking changes can happen in minor versions.
 ## 0.1
 
 This is the first public release of PickerKit.
+
+### ✨ Features
+
+* `Camera` is a camera-based image picker.
+* `DocumentScanner` is a Vision-based document scanner.
+* `ImagePicker` is a UIKit image picker wrapper.